### 2025-07-30 #8669 @editorialbot
[PRE REVIEW]: NAMOSIM: a Robot Motion Planner for Navigation Among Movable Obstacles
**Submitting author:** @david-wb (<a href="http://orcid.org/0009-0000-3804-7734">David Brown</a>)
**Repository:** https://github.com/Chroma-CITI/namosim
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.0
