### 2025-07-30 #8669 @crvernon
:wave: @david-wb - we are managing a large queue of submissions right now so I'll have to waitlist this submission until I can get a topic editor assigned for you.

In the meantime, please reduce the word count of your paper to <=1000 words.
