### 2025-08-14 #8816 @sea-bass
@boisgera seems you need to assign us to the git issue so we can use editorialbot commands. Thanks!
