### 2025-08-19 Chroma-CITI/namosim#3 @SteveMacenski
JOSS Review: Instructions
After cloning, tell people to cd into the directory.

Also, fix spelling for "Alternativley"
