### 2025-08-19 Chroma-CITI/namosim#4 @SteveMacenski
JOSS Review: Show Automated Testing
The review requires automated testing. I only see 1 PR which doesn't have successful software build / testing. Please add and link me a job showing that it properly builds and runs the automated test suite
