### 2025-08-19 Chroma-CITI/namosim#5 @SteveMacenski
JOSS Review: Functionality
The test only shows 2 robots moving around each other, but not an example of movable objects being moved by the robots. Please update the testing demo so that it shows something like what the paper is outlining.
