### 2025-08-19 Chroma-CITI/namosim#3 @SteveMacenski
Does this work require Cuda? I see the requirements.txt pulling in alot of cuda packages that are not outlined in the system requirements in the readme or documentation. That is a key item. If it is not required, please reduce the requirements list to be only containing what is required.
