### 2025-08-19 Chroma-CITI/namosim#6 @SteveMacenski
JOSS Review: Documentation
I think the documentation in the readme/website needs to expand on:
- The statement of need
- Example usage + algorithms provided and how to use each + how to work with it to develop additional NAMO agents (since that is the crux of this work)
- Community contribution guidelines
- Testing requirements and methods
