### 2025-08-19 #8816 @SteveMacenski
I have some topics to address with the authors:
- https://github.com/Chroma-CITI/namosim/issues/2 --> authorship seems like it has extra authors than folks that have been involved in the project
- The paper needs to adjust "ROS2" to "ROS 2" as the proper name and cite its source as a key point of its integration contribution story.
- The paper should cite the method used for collision detection they describe as convex swept volume
-  https://github.com/Chroma-CITI/namosim/issues/3 --> Issues with the instructions (though very minor). Also possible missing requirement for cuda or pulling in excess dependencies?
- https://github.com/Chroma-CITI/namosim/issues/4 --> Automated testing / example needed
- https://github.com/Chroma-CITI/namosim/issues/5 --> Functionality: the demo provided in the documentation does not show an example of the contribution laid out in the paper
- https://github.com/Chroma-CITI/namosim/issues/6 --> Documentation: I believe this is lacking substantially by the requirements of this journal currently, but they have a good start.

@boisgera the test coverage covers only 27% of the total lines. I don't know if there's a quality bar for automated testing of this repository, but that seems a bit low for the *Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?* line item. How should I interpret that?
