### 2025-08-23 Chroma-CITI/namosim#10 @sea-bass
[JOSS Review] Paper feedback
This is my overall feedback on the paper submission. I will probably add more as I'm able to run the code and play around.

## General

- [x] Overall, this paper needs citations to other work. As it stands now, 3 out of 4 citations are of the author's own work, and the last one is an algorithm that this tool implements. Citations for JOSS don't need to include a thorough literature review, by any means, but should definitely cite other competitive/complementary software tools that exist out there.
- [x] As the other reviewer already pointed out, please replace all uses of ROS2 with "ROS 2" -- not just in the paper, but in the software and docs as well.
- [x] An architecture diagram would be a nice accompanying figure, so that the paper isn't all text besides Figure 1, and more importantly because it would help explain how the tool works!
- [x] I notice from the code that there are other interesting agents such as RRT / PPO / etc. Worth mentioning in the paper that you can add your own agents, citing these as examples?

## Statement of Need

- [ ] I think this section is reasonable text-wise, but it needs to reference other software tools that could be used to simulate NAMO scenarios.

https://github.com/openjournals/joss-reviews/issues/8816
