### 2025-08-23 Chroma-CITI/namosim#3 @sea-bass
> Does this work require Cuda?

Was wondering the same thing -- it makes the devcontainer build super slow, and if it's not required I'd consider removing it.
