### 2025-08-25 8116afd
remove dependency on pytorch
