+++ namosim/agents/ppo_agent.py
@@ -1,586 +1,0 @@
-# pyright: reportConstantRedefinition=false
-
-import copy
-import math
-import random
-import typing as t
-
-import numpy as np
-import numpy.typing as npt
-import torch
-import torchvision.transforms as transforms
-from PIL import Image
-from shapely.geometry import Point, Polygon
-from shapely import affinity
-from transformers.models.auto.image_processing_auto import AutoImageProcessor
-from typing_extensions import Self
-
-import namosim.display.ros2_publisher as rp
-import namosim.navigation.action_result as ar
-import namosim.navigation.basic_actions as ba
-from namosim.svg_styles import AgentStyle
-import namosim.world.world as w
-from namosim.agents.agent import Agent, RLThinkResult
-from namosim.agents.models import DEIT_MODEL_CHECKPOINT, PPOActor
-from namosim.algorithms import graph_search
-from namosim.data_models import GridCellModel, Pose2D, PPOAgentConfigModel
-from namosim.input import Input
-from namosim.log import logger
-from namosim.utils import utils
-from namosim.world.binary_occupancy_grid import (
-    BinaryOccupancyGrid,
-)
-from namosim.world.entity import Movability
-from namosim.world.goal import Goal
-from namosim.world.obstacle import Obstacle
-from namosim.world.sensors.omniscient_sensor import OmniscientSensor
-
-
-class State:
-    def __init__(
-        self,
-        *,
-        grid: npt.NDArray[np.float32],
-        goal_state_grid: npt.NDArray[np.float32],
-        robot_pose: Pose2D,
-        normalized_robot_pose: npt.NDArray[np.float32],
-        goal_pose: npt.NDArray[np.float32],
-        normalized_goal_pose: npt.NDArray[np.float32],
-        goal_reached: bool,
-    ):
-        self.grid = grid
-        self.goal_state_grid = goal_state_grid
-        self.robot_pose = robot_pose
-        self.normalized_robot_pose = normalized_robot_pose
-        self.goal_pose = goal_pose
-        self.normalized_goal_pose = normalized_goal_pose
-        self.goal_reached = goal_reached
-
-
-class RLAgentStepResult:
-    def __init__(
-        self,
-        *,
-        state: State,
-        goal_reached: bool,
-        action: ba.Action,
-        action_idx: int,
-        action_log_prob: float,
-    ):
-        self.state = state
-        self.goal_reached = goal_reached
-        self.action_idx = action_idx
-        self.action_log_prob = action_log_prob
-        self.action = action
-
-
-class PPOAgent(Agent):
-    def __init__(
-        self,
-        *,
-        navigation_goals: t.List[Goal],
-        config: PPOAgentConfigModel,
-        logs_dir: str,
-        uid: str,
-        polygon: Polygon,
-        pose: Pose2D,
-        sensors: t.List[OmniscientSensor],
-        logger: utils.NamosimLogger,
-        cell_size: float,
-        style: AgentStyle | None = None,
-    ):
-        Agent.__init__(
-            self,
-            uid=uid,
-            navigation_goals=navigation_goals,
-            config=config,
-            logs_dir=logs_dir,
-            polygon=polygon,
-            pose=pose,
-            sensors=sensors,  # type: ignore
-            style=style,
-            logger=logger,
-            cell_size=cell_size,
-        )
-        self.config = config
-        self.robot_max_inflation_radius = utils.get_circumscribed_radius(self.polygon)
-        self.steps_per_goal = 0
-        self.step_count = 0
-        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        self.actor: PPOActor | None = None
-        self.dist_to_goal: t.Dict[GridCellModel, float] = {}
[…]
