+++ namosim/agents/models.py
@@ -1,244 +1,0 @@
-import torch
-import torch.nn as nn
-from transformers.models.deit import DeiTModel
-
-DEIT_MODEL_CHECKPOINT = "facebook/deit-tiny-distilled-patch16-224"
-
-
-class PPOBaseModel(nn.Module):
-    def __init__(self):
-        super(PPOBaseModel, self).__init__()
-        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        self.vit: DeiTModel = DeiTModel.from_pretrained(DEIT_MODEL_CHECKPOINT)  # type: ignore
-        self.vit.to(self.device)  # type: ignore
-        self.vit.train()  # type: ignore
-
-    def forward(self, imgs: torch.Tensor):  # type: ignore
-        last_hidden_state = self.vit(imgs).last_hidden_state
-        return last_hidden_state[:, 0]
-
-
-class SimpleCNN(nn.Module):
-    def __init__(self, action_size: int):
-        super().__init__()
-        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        self.conv1 = nn.Conv2d(1, 16, kernel_size=3, stride=1, padding=1)
-        self.relu1 = nn.ReLU()
-        self.pool1 = nn.MaxPool2d(kernel_size=2, stride=2)
-        self.conv2 = nn.Conv2d(16, 32, kernel_size=3, stride=1, padding=1)
-        self.relu2 = nn.ReLU()
-        self.pool2 = nn.MaxPool2d(kernel_size=2, stride=2)
-        self.fc = nn.Linear(32 * 56 * 56, 128)
-        self.relu3 = nn.ReLU()
-        self.out = nn.Linear(128, action_size)
-
-    def forward(self, imgs: torch.Tensor):
-        x = self.conv1(imgs)
-        x = self.relu1(x)
-        x = self.pool1(x)
-        x = self.conv2(x)
-        x = self.relu2(x)
-        x = self.pool2(x)
-        x = x.view(x.size(0), -1)
-        x = self.fc(x)
-        x = self.relu3(x)
-        x = self.out(x)
-        x = nn.functional.softmax(x, dim=-1)
-        return x
-
-
-class SSAgentModel(nn.Module):
-    def __init__(self, embedding_size: int = 512, action_size: int = 7):
-        super().__init__()
-        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        # self.start_state_encoder = nn.Sequential(
-        #     nn.Conv2d(1, 16, kernel_size=3, stride=2, padding=1),  # 112x112
-        #     nn.ReLU(True),
-        #     nn.Conv2d(16, 32, kernel_size=3, stride=2, padding=1),  # 56x56
-        #     nn.ReLU(True),
-        #     nn.Conv2d(32, 64, kernel_size=3, stride=2, padding=1),  # 28x28
-        #     nn.ReLU(True),
-        #     nn.Conv2d(64, 128, kernel_size=3, stride=2, padding=1),  # 14x14
-        #     nn.ReLU(True),
-        #     nn.Conv2d(
-        #         128, embedding_size, kernel_size=7, stride=1, padding=0
-        #     ),  # 8x8 (output size can be adjusted by kernel size)
-        #     nn.Flatten(),
-        #     nn.Linear(8 * 8 * embedding_size, embedding_size),
-        # )
-        # self.goal_state_encoder = nn.Sequential(
-        #     nn.Conv2d(1, 16, kernel_size=3, stride=2, padding=1),  # 112x112
-        #     nn.ReLU(True),
-        #     nn.Conv2d(16, 32, kernel_size=3, stride=2, padding=1),  # 56x56
-        #     nn.ReLU(True),
[…]
