+++ namosim/utils/conversion.py
@@ -17,8 +17,8 @@ from shapely.geometry.base import BaseGeometry
 from shapely.ops import polygonize, unary_union
 from svgpath2mpl import parse_path
 import triangle
-
-from namosim.display.conversions import polygon_to_triangle_vertices
+from namosim.log import logger
+from shapely.validation import make_valid
 
 SVG_PATH_ATTRIBUTES_WHITELIST = ["id", "d", "style", "type"]
 
@@ -259,6 +259,146 @@ def rgb_tuple_to_hex(rgb):
     return "#{0:02x}{1:02x}{2:02x}".format(
         color_clamp(rgb[0]), color_clamp(rgb[1]), color_clamp(rgb[2])
     )
+
+
+def polygon_to_triangle_vertices(shapely_geometry):
+    """
+    Convert a Shapely Polygon or MultiPolygon to a list of triangle vertices.
+    Returns a list of triangles, where each triangle is a list of 3 vertex coordinates.
+    """
+    # Handle MultiPolygon by processing each Polygon separately
+    if isinstance(shapely_geometry, MultiPolygon):
+        logger.debug(
+            "Processing MultiPolygon with %d polygons", len(shapely_geometry.geoms)
+        )
+        all_triangles = []
+        for poly in shapely_geometry.geoms:
+            # Recursively call for each Polygon in the MultiPolygon
+            triangles = polygon_to_triangle_vertices(poly)
+            all_triangles.extend(triangles)
+        return all_triangles
+
+    # Ensure the input is a Polygon
+    if not isinstance(shapely_geometry, Polygon):
+        logger.error(
+            "Input geometry is not a Polygon or MultiPolygon: %s",
+            type(shapely_geometry),
+        )
+        return []
+
+    # Check if the polygon is valid
+    if not shapely_geometry.is_valid:
+        logger.warning("Invalid polygon detected. Attempting to repair.")
+        try:
+            shapely_geometry = make_valid(shapely_geometry)
+            # make_valid may return a MultiPolygon, so recurse
+            if isinstance(shapely_geometry, MultiPolygon):
+                logger.debug("Repaired geometry is a MultiPolygon")
+                return polygon_to_triangle_vertices(shapely_geometry)
+            if not shapely_geometry.is_valid:
+                logger.error("Polygon repair failed. Skipping triangulation.")
+                return []
+        except Exception as e:
+            logger.error(f"Error repairing polygon: {e}")
+            return []
+
+    # Simplify the polygon to remove near-degenerate edges
+    shapely_geometry = t.cast(
+        Polygon, shapely_geometry.simplify(tolerance=1e-5, preserve_topology=True)
+    )
+
+    # Check for near-zero area
+    if shapely_geometry.area < 1e-10:
+        logger.warning("Polygon has near-zero area. Skipping triangulation.")
+        return []
+
+    # Extract exterior coordinates (excluding the last point, which repeats the first)
+    exterior_coords = np.array(shapely_geometry.exterior.coords)[:-1]
+    if len(exterior_coords) < 3:
+        logger.warning("Exterior ring has too few points. Skipping triangulation.")
+        return []
+
+    # Initialize vertices with exterior coordinates
+    vertices = exterior_coords.copy()
+
+    # Create segments for the exterior
+    segments = np.array(
+        [[i, (i + 1) % len(exterior_coords)] for i in range(len(exterior_coords))]
+    )
+
+    # Handle holes if the polygon has any
+    holes = []
+    if shapely_geometry.interiors:
+        for interior in shapely_geometry.interiors:
+            # Extract interior coordinates (excluding the last point)
+            interior_coords = np.array(interior.coords)[:-1]
+            if len(interior_coords) < 3:
+                logger.warning("Interior ring has too few points. Skipping hole.")
+                continue
[…]
