+++ namosim/display/conversions.py
@@ -9,6 +9,7 @@ from builtin_interfaces.msg import Time
 from geometry_msgs.msg import Point, Pose, Quaternion, Vector3
 from grid_map_msgs.msg import GridMap
 from shapely.geometry import Polygon, MultiPolygon
+from namosim.utils.conversion import polygon_to_triangle_vertices
 from namosim.world.binary_occupancy_grid import BinaryOccupancyGrid
 from namosim.world.entity import Style
 from std_msgs.msg import (
@@ -29,8 +30,6 @@ from namosim.display import tf_replacement
 from namosim.navigation.path_type import PathType
 from namosim.utils import utils
 import triangle
-from shapely.validation import make_valid
-from namosim.log import logger
 
 
 def plan_to_markerarray(
@@ -115,146 +114,6 @@ def real_path_to_linestrip(
         marker.points.append(Point(x=real_path[-1][0], y=real_path[-1][1], z=z_index))  # type: ignore
         marker.points.append(Point(x=link_point[0], y=link_point[1], z=z_index))  # type: ignore
     return marker
-
-
-def polygon_to_triangle_vertices(shapely_geometry):
-    """
-    Convert a Shapely Polygon or MultiPolygon to a list of triangle vertices.
-    Returns a list of triangles, where each triangle is a list of 3 vertex coordinates.
-    """
-    # Handle MultiPolygon by processing each Polygon separately
-    if isinstance(shapely_geometry, MultiPolygon):
-        logger.debug(
-            "Processing MultiPolygon with %d polygons", len(shapely_geometry.geoms)
-        )
-        all_triangles = []
-        for poly in shapely_geometry.geoms:
-            # Recursively call for each Polygon in the MultiPolygon
-            triangles = polygon_to_triangle_vertices(poly)
-            all_triangles.extend(triangles)
-        return all_triangles
-
-    # Ensure the input is a Polygon
-    if not isinstance(shapely_geometry, Polygon):
-        logger.error(
-            "Input geometry is not a Polygon or MultiPolygon: %s",
-            type(shapely_geometry),
-        )
-        return []
-
-    # Check if the polygon is valid
-    if not shapely_geometry.is_valid:
-        logger.warning("Invalid polygon detected. Attempting to repair.")
-        try:
-            shapely_geometry = make_valid(shapely_geometry)
-            # make_valid may return a MultiPolygon, so recurse
-            if isinstance(shapely_geometry, MultiPolygon):
-                logger.debug("Repaired geometry is a MultiPolygon")
-                return polygon_to_triangle_vertices(shapely_geometry)
-            if not shapely_geometry.is_valid:
-                logger.error("Polygon repair failed. Skipping triangulation.")
-                return []
-        except Exception as e:
-            logger.error(f"Error repairing polygon: {e}")
-            return []
-
-    # Simplify the polygon to remove near-degenerate edges
-    shapely_geometry = t.cast(
-        Polygon, shapely_geometry.simplify(tolerance=1e-5, preserve_topology=True)
-    )
-
-    # Check for near-zero area
-    if shapely_geometry.area < 1e-10:
-        logger.warning("Polygon has near-zero area. Skipping triangulation.")
-        return []
-
-    # Extract exterior coordinates (excluding the last point, which repeats the first)
-    exterior_coords = np.array(shapely_geometry.exterior.coords)[:-1]
-    if len(exterior_coords) < 3:
-        logger.warning("Exterior ring has too few points. Skipping triangulation.")
-        return []
-
-    # Initialize vertices with exterior coordinates
-    vertices = exterior_coords.copy()
-
-    # Create segments for the exterior
-    segments = np.array(
[…]
