### 2025-08-25 Chroma-CITI/namosim#3 @david-wb
The dependency on cuda (via pytorch) has been removed (#11).
I added brief instructions on how to launch a scenario via the CLI instead of pytest. (#12).
