+++ CHANGELOG.rst
@@ -1,0 +1,1641 @@
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Changelog for package namosim
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Forthcoming
+-----------
+* Merge pull request `#12 <https://github.com/Chroma-CITI/namosim/issues/12>`_ from Chroma-CITI/rm_torch
+  Update instructions
+* docs
+* Merge pull request `#11 <https://github.com/Chroma-CITI/namosim/issues/11>`_ from Chroma-CITI/rm_torch
+  remove dependency on pytorch
+* typo
+* fix circular import
+* remove imports of ppo agent
+* remove imports of ppo agent
+* remove imports of ppo agent
+* remove dependency on pytorch
+* add jacques as maintainer
+* ci
+* ci
+* ci
+* skip unit tests for colcon test
+* rm pip deps from package.xml
+* move pip packages to exec_depend
+* ci
+* ci
+* show colcon test output
+* remove dependency on aabbtree
+* fix dependencies and colcon test
+* add draft-pdf workflow
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* docs
+* add orcid ids
+* docs
+* fix citations
+* docs
+* docs
+* docs
+* docs and ci
+* ci
+* docs
+* ci
+* update github action
+* Merge branch 'humble' of github.com:Chroma-CITI/namosim into humble
+* update github action
+* Merge pull request `#1 <https://github.com/Chroma-CITI/namosim/issues/1>`_ from Chroma-CITI/paper
+  Paper
+* update github action
+* add github action
+* add github action
+* paper
+* paper
+* paper
+* paper
+* paper
+* paper
+* paper
+* paper
+* paper
+* paper
+* paper
+* paper
+* paper
+* Merge branch 'joss' into 'humble'
+  add joss paper
+  See merge request chroma/namo/namosim!196
+* add joss paper
+* Merge branch 'three_robots' into 'humble'
+  Three robots
+  See merge request chroma/namo/namosim!194
+* use absolute cost values when compuing evasion cost
+* is_for_deadlock
+* fix tests
+* format
+* types
+* fix types
+* clean up
+* clean up
+* clean up
+* simplify conflict detection
+* simplifying conflict detection
+* clean up
+* clean up
+* remove unnecessary aabb trees
+* simplifying conflict detection
+* Merge branch 'stilman_rrt' into 'humble'
+  stilman rrt
+  See merge request chroma/namo/namosim!191
+* PoseModel -> Pose2D
+* use named tuple for PoseModel
+* fix set_current_action_index
+* fix set_current_action_index
+* add function to set current action index on navigation plan
+* fix tests
+* fix stealing movable test
+* fix tests
+* test kd tree
+* use obstacle pose not robot pose for social cost
+* penalize rotations
+* stilman rrt
+* add collision margin for robot_obstacle_polyon in RRT* agent
+* add minim discretization of poses in RRT tree
+* cache poses for rrt star
+* collision bug fix
+* fair comparison
+* stilman rrt
+* default to diff drive
+* stilman rrt
+* stilman rrt
+* Merge branch 'humble' into stilman_rrt
+* Merge branch 'triangles' into 'humble'
+  fix polygon triangulation bug
+  See merge request chroma/namo/namosim!192
+* propagate non-zero exit codes in tests
+* fix polygon triangulation bug
+* use correct center or rotation in rrt
+* stilman rrt
+* Merge branch 'minor' into 'humble'
[…]
