+++ setup.py
@@ -13,7 +13,7 @@ def read_requirements():
 
 setup(
     name=package_name,
-    version="0.0.0",
+    version="0.0.1",
     packages=find_packages(
         exclude=["tests", "tests.*"]
     ),  # Exclude tests and subpackages
