### 2025-08-26 Chroma-CITI/namosim#4 @david-wb
Hi @SteveMacenski. There's a github ci badge on the main readme. If you click that you can see the build logs. Here's a link to one.
https://github.com/Chroma-CITI/namosim/actions/runs/17234620512/job/48896469629
Is that what you are referring to?
