+++ namosim/config.py
@@ -1,4 +1,7 @@
 import os
 
 DISPLAY_WINDOW = os.environ.get("NAMO_NO_DISPLAY_WINDOW", "") == ""
-DEACTIVATE_RVIZ = os.environ.get("NAMO_DEACTIVATE_RVIZ", "") != ""
+DEACTIVATE_RVIZ = (
+    os.environ.get("NAMO_DEACTIVATE_RVIZ", "") != ""
+    or os.environ.get("ROS_DISTRO", "") == ""
+)
