### 2025-08-26 3a08b29
move docs from readme to docs site
