+++ _sources/guides.rst.txt
@@ -1,3 +1,5 @@
+.. _guides:
+
 Guides
 =============
 
