### 2025-08-27 Chroma-CITI/namosim#5 @david-wb
I think this is done. The first example on the [usage page](https://chroma-citi.github.io/namosim/usage.html) is a movable obstacle scenario.
