### 2025-08-27 426a324
add a guide for creating custom agents
