+++ docs/source/creating_custom_scenarios.rst
@@ -1,0 +1,10 @@
+Creating Custom Scenarios
+========================
+
+The NAMO planner supports loading scenarios in one of two ways:
+
+1. From an SVG file
+2. From a ROS map
+
+The first method is best when executing purely in simulation, while the second is intended for usage with
+real robots.

[…]
