### 2025-08-27 919dbe2
add contributing guidelines
