+++ docs/source/index.rst
@@ -18,27 +18,46 @@ obstacles may be grasped and moved. This problem is pertinent for real-world rob
 where robots may need to move or manipulate objects in order to navigate and complete their tasks.
 
 Cite Us
-***********
-If you use NAMO in any capacity for your research, please do cite the associated paper:
-::
-  @inproceedings{renault:hal-02912925,
-    TITLE = {{Modeling a Social Placement Cost to Extend Navigation Among Movable Obstacles (NAMO) Algorithms}},
-    AUTHOR = {Renault, Benoit and Saraydaryan, Jacques and Simonin, Olivier},
-    URL = {https://hal.archives-ouvertes.fr/hal-02912925},
-    BOOKTITLE = {{IROS 2020 - IEEE/RSJ International Conference on Intelligent Robots and Systems}},
-    ADDRESS = {Las Vegas, United States},
-    SERIES = {2020 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS) Conference Proceedings},
-    PAGES = {11345-11351},
-    YEAR = {2020},
-    MONTH = Oct,
-    DOI = {10.1109/IROS45743.2020.9340892},
-    KEYWORDS = {Navigation Among Movable Obstacles (NAMO) ; Socially- Aware Navigation (SAN) ; Path planning ; Simulation},
-    PDF = {https://hal.archives-ouvertes.fr/hal-02912925/file/IROS_2020_Camera_Ready.pdf},
-    HAL_ID = {hal-02912925},
-    HAL_VERSION = {v1},
-  }
+=======
 
-Check out our IROS24 Paper at `https://hal.science/hal-04705395 <https://hal.science/hal-04705395>`_.
+If you reuse any part of this project in your research, please cite the associated papers:
+
+.. code-block:: bibtex
+
+    @inproceedings{renault:hal-04705395,
+      TITLE = {{Multi-Robot Navigation among Movable Obstacles: Implicit Coordination to Deal with Conflicts and Deadlocks}},
+      AUTHOR = {Renault, Benoit and Saraydaryan, Jacques and Brown, David and Simonin, Olivier},
+      URL = {https://hal.science/hal-04705395},
+      BOOKTITLE = {{IROS 2024 - IEEE/RSJ International Conference on Intelligent Robots and Systems}},
+      ADDRESS = {Abu DHABI, United Arab Emirates},
+      PUBLISHER = {{IEEE}},
+      PAGES = {1-7},
+      YEAR = {2024},
+      MONTH = Oct,
+      KEYWORDS = {Planning ; Scheduling and Coordination ; Path Planning for Multiple Mobile Robots or Agents ; Multi-Robot Systems},
+      PDF = {https://hal.science/hal-04705395v1/file/IROS24_1134_FI.pdf},
+      HAL_ID = {hal-04705395},
+      HAL_VERSION = {v1},
+    }
+
+.. code-block:: bibtex
+
+    @inproceedings{renault:hal-02912925,
+      TITLE = {{Modeling a Social Placement Cost to Extend Navigation Among Movable Obstacles (NAMO) Algorithms}},
+      AUTHOR = {Renault, Benoit and Saraydaryan, Jacques and Simonin, Olivier},
+      URL = {https://hal.archives-ouvertes.fr/hal-02912925},
+      BOOKTITLE = {{IROS 2020 - IEEE/RSJ International Conference on Intelligent Robots and Systems}},
+      ADDRESS = {Las Vegas, United States},
+      SERIES = {2020 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS) Conference Proceedings},
[…]
