+++ _sources/contributing.rst.txt
@@ -1,0 +1,196 @@
+Contributing to NAMOSIM
+======================
+
+Thank you for your interest in contributing to NAMOSIM! We welcome contributions from the community to improve the project, whether through code, documentation, bug reports, or feature suggestions. This page outlines the guidelines for contributing, with a specific focus on testing requirements and methods to ensure the quality and reliability of the NAMOSIM codebase.
+
+How to Contribute
+-----------------
+
+1. **Fork the Repository**: Start by forking the NAMOSIM repository on GitHub to your own account.
+2. **Clone the Fork**: Clone your forked repository to your local machine.
+3. **Create a Branch**: Create a new branch for your changes (e.g., `feature/new-planner` or `bugfix/issue-123`).
+4. **Make Changes**: Implement your changes, ensuring they align with the project's coding standards (see below).
+5. **Test Your Changes**: Run the test suite and add new tests as needed (see Testing Requirements below).
+6. **Submit a Pull Request**: Push your changes to your fork and submit a pull request to the main NAMOSIM repository. Include a clear description of your changes and reference any related issues.
+
+Please ensure your contributions adhere to the following guidelines:
+
+- Follow the existing code style (PEP 8 for Python code).
+- Write clear, concise commit messages.
+- Update documentation if your changes affect usage or functionality.
+- Ensure all tests pass before submitting a pull request.
+
+Testing Requirements
+-------------------
+
+To maintain the reliability and correctness of NAMOSIM, all code contributions must include appropriate tests. Testing ensures that new features or bug fixes do not introduce regressions and that the planner behaves as expected in various scenarios. NAMOSIM is built and tested in a ROS Humble environment using `colcon` and custom test scripts.
+
+### Testing Environment
+NAMOSIM uses the `osrf/ros:humble-desktop-full` Docker container for its CI pipeline, which includes ROS Humble and dependencies. To replicate this environment locally:
+
+- **Use ROS Humble**: Ensure you have ROS Humble installed. Refer to `installation.rst` for setup instructions.
+- **Install Dependencies**: Install required system packages and Python dependencies:
+  .. code-block:: bash
+
+      sudo apt-get update
+      sudo apt-get install -y swig curl python3-pip ros-humble-grid-map-msgs
+      rosdep install --from-paths . -r -y
+      pip install -r requirements.txt
+
+- **Set Up ROS Environment**: Source the ROS Humble setup script before building or testing:
+  .. code-block:: bash
+
+      source /opt/ros/humble/setup.bash
+
+### Types of Tests
+NAMOSIM uses a combination of the following tests to validate functionality:
+
+- **Linting/Formatting**: Ensures code adheres to style guidelines (e.g., PEP 8) using tools like `flake8` or `black`.
+- **Type Checking**: Verifies type correctness in Python code using static type checkers like `mypy`.
+- **Unit Tests**: Test individual components (e.g., path planning algorithms, obstacle manipulation logic) in isolation. These should cover edge cases and common use cases.
+- **End-to-End Tests**: Verify the entire system, from environment setup to path planning and obstacle manipulation, in realistic scenarios.
+
+### Testing Framework
+NAMOSIM uses `colcon` for building and testing, with tests executed via custom scripts (`scripts/test_unit.sh`, `scripts/test_e2e.sh`, `scripts/test_types.sh`). Tests are organized in the `tests/` directory and should be compatible with the ROS Humble environment.
+
+### Writing Tests
+When adding new functionality or fixing bugs, include corresponding tests in the `tests/` directory. Follow these guidelines:
+
[…]
