+++ docs/source/contributing.rst
@@ -23,7 +23,7 @@ Please ensure your contributions adhere to the following guidelines:
 Testing Requirements
 -------------------
 
-To maintain the reliability and correctness of NAMOSIM, all code contributions must include appropriate tests. Testing ensures that new features or bug fixes do not introduce regressions and that the planner behaves as expected in various scenarios. NAMOSIM is built and tested in a ROS Humble environment using `colcon` and custom test scripts.
+To maintain the reliability and correctness of NAMOSIM, all code contributions must include appropriate tests. Testing ensures that new features or bug fixes do not introduce regressions and that the planner behaves as expected in various scenarios.
 
 ### Testing Environment
 NAMOSIM uses the `osrf/ros:humble-desktop-full` Docker container for its CI pipeline, which includes ROS Humble and dependencies. To replicate this environment locally:
@@ -45,111 +45,79 @@ NAMOSIM uses the `osrf/ros:humble-desktop-full` Docker container for its CI pipe
 ### Types of Tests
 NAMOSIM uses a combination of the following tests to validate functionality:
 
-- **Linting/Formatting**: Ensures code adheres to style guidelines (e.g., PEP 8) using tools like `flake8` or `black`.
-- **Type Checking**: Verifies type correctness in Python code using static type checkers like `mypy`.
-- **Unit Tests**: Test individual components (e.g., path planning algorithms, obstacle manipulation logic) in isolation. These should cover edge cases and common use cases.
+- **Linting/Formatting**: Ensures code adheres to style guidelines using `black`.
+- **Type Checking**: Verifies type correctness in Python code using the static type checkers `pyright`.
+- **Unit Tests**: Test individual components (e.g., path planning algorithms, obstacle manipulation logic) in isolation. These should run fast and cover edge cases and common use cases.
 - **End-to-End Tests**: Verify the entire system, from environment setup to path planning and obstacle manipulation, in realistic scenarios.
 
 ### Testing Framework
-NAMOSIM uses `colcon` for building and testing, with tests executed via custom scripts (`scripts/test_unit.sh`, `scripts/test_e2e.sh`, `scripts/test_types.sh`). Tests are organized in the `tests/` directory and should be compatible with the ROS Humble environment.
+NAMOSIM tests are executed via custom scripts (`scripts/test_unit.sh`, `scripts/test_e2e.sh`, `scripts/test_types.sh`). Tests are organized in the `tests/` directory.
 
 ### Writing Tests
 When adding new functionality or fixing bugs, include corresponding tests in the `tests/` directory. Follow these guidelines:
 
-- **Test Structure**: Place tests in the `tests/` directory, organized by module (e.g., `tests/test_planner.py` for unit tests, `tests/e2e/e2e_test.py` for end-to-end tests).
-- **Test Naming**: Use descriptive names for test functions, prefixed with `test_` (e.g., `test_path_planner_handles_collisions`).
-- **Coverage**: Aim for high test coverage, especially for critical components like path planning and obstacle manipulation. Use `colcon test` to verify functionality.
-- **Edge Cases**: Include tests for edge cases, such as empty environments, fully blocked paths, or maximum obstacle weight limits.
+- **Test Structure**: Place tests in the `tests/` directory, organized by module (e.g., `tests/planner_test.py` for unit tests, `tests/e2e/e2e_test.py` for end-to-end tests).
+- **Test Naming**: Use descriptive names for test files, post-fixed with `_test.py` (e.g., `collision_test.py`).
[…]
