### 2025-08-27 Chroma-CITI/namosim#6 @david-wb
These points have now been addressed on the docs site with:
* an adjusted [statement of need](https://chroma-citi.github.io/namosim/index.html#statement-of-need) (now included in the docs)
* a new guide page on [how to create custom agents](https://chroma-citi.github.io/namosim/creating_custom_agents.html)
* a community [contributions guidelines page](https://chroma-citi.github.io/namosim/contributing.html) including testing requirements
* a page on [testing methods](https://chroma-citi.github.io/namosim/testing.html)
