+++ namosim/agents/agent.py
@@ -25,6 +25,12 @@ import numpy as np
 
 
 class ThinkResult:
+    """Represents the result of an agent's thinking process.
+
+    This class encapsulates the outcome of an agent's decision-making cycle,
+    including planning information, goal details, and any conflicts detected.
+    """
+
     def __init__(
         self,
         *,
@@ -48,6 +54,12 @@ class ThinkResult:
 
 
 class RLThinkResult(ThinkResult):
+    """Represents the result of a reinforcement learning agent's thinking process.
+
+    Extends ThinkResult with additional information specific to RL agents,
+    such as log probability and action index.
+    """
+
     def __init__(
         self,
         *,
@@ -82,6 +94,12 @@ class RLThinkResult(ThinkResult):
 
 
 class Agent(Entity):
+    """Represents an autonomous agent in the simulation environment.
+
+    An agent is capable of sensing its environment, planning actions,
+    and executing movements towards goals while handling conflicts.
+    """
+
     def __init__(
         self,
         *,
@@ -97,7 +115,21 @@ class Agent(Entity):
         logger: utils.NamosimLogger,
         style: svg_styles.AgentStyle | None = None,
     ):
-        self.agent_style = style if style else svg_styles.AgentStyle()
+        """Initialize the Agent.
+
+        Args:
+            config: Configuration object for the agent.
+            navigation_goals: List of goals for the agent to navigate to.
+            logs_dir: Directory path for logging agent activities.
+            uid: Unique identifier for the agent.
+            polygon: Shapely Polygon representing the agent's physical shape.
+            pose: Current pose of the agent (x, y, theta).
+            sensors: List of sensors attached to the agent.
+            cell_size: Size of the grid cells for path planning.
+            movability: Movability status of the agent.
+            logger: Logger instance for recording events.
+            style: Optional style configuration for visualization.
+        """
         super().__init__(
             polygon=polygon,
             pose=pose,
@@ -113,7 +145,7 @@ class Agent(Entity):
         self.logger = logger
         self._world: t.Optional["w.World"] = None
         self._navigation_goals = navigation_goals
-
+        self.agent_style = style if style else svg_styles.AgentStyle()
         self.logs_dir = logs_dir
 
         self.__last_action_result: ActionResult | None = None
@@ -137,6 +169,11 @@ class Agent(Entity):
         """
 
     def copy(self) -> "Agent":
+        """Create a deep copy of the agent.
+
+        Returns:
+            A new Agent instance with copied attributes.
+        """
         return Agent(
             config=self.config,
             navigation_goals=copy.deepcopy(self._navigation_goals),
@@ -152,16 +189,33 @@ class Agent(Entity):
         )
 
     def set_navigation_goals(self, goals: t.List[goal.Goal]):
+        """Set the list of navigation goals for the agent.
+
+        Args:
+            goals: List of Goal objects to navigate to.
+        """
         self._navigation_goals = goals
         self._goal = None
 
     def init(self, world: "w.World") -> None:
+        """Initialize the agent with the world state.
+
+        Args:
+            world: The world object containing the environment.
+        """
         self._world = world.light_copy([])
         self.is_initialized = True
 
     def sense(
         self, ref_world: "w.World", last_action_result: ActionResult, step_count: int
     ):
+        """Update the agent's perception of the world using sensors.
+
+        Args:
+            ref_world: Reference world state.
+            last_action_result: Result of the last action performed.
+            step_count: Current simulation step count.
+        """
         self._last_action_result = last_action_result
         (
             self._added_uids,
@@ -176,6 +230,15 @@ class Agent(Entity):
         ros_publisher: t.Optional["rp.RosPublisher"] = None,
[…]
