### 2025-08-28 c099413
add architecture diagram
