+++ scripts/test_unit.sh
@@ -5,5 +5,5 @@ set -eo pipefail
 DIR=$(dirname "$0")
 cd $DIR/..
 
-export NAMO_NO_DISPLAY_WINDOW="TRUE"
+export NAMO_DEACTIVATE_TKINTER="TRUE"
 python3 -m pytest -s --cov=namosim --cov-fail-under=27 tests/unit
