+++ scripts/test_e2e.sh
@@ -5,6 +5,6 @@ set -eo pipefail
 DIR=$(dirname "$0")
 cd $DIR/..
 
-export NAMO_NO_DISPLAY_WINDOW="TRUE"
+export NAMO_DEACTIVATE_TKINTER="TRUE"
 export NAMO_DEACTIVATE_RVIZ="TRUE"
 python3 -m pytest -s --cov=namosim --cov-fail-under=55 tests/e2e
