### 2025-08-28 Chroma-CITI/namosim#10 @david-wb
Hi @sea-bass, thanks for your comments. I’ve added an architecture diagram, which you can see in the paper. The PPO agent was incomplete and had heavy dependencies, so I removed it. The RRT agent is a modified Stilman agent that uses RRT to accelerate the obstacle manipulation search. Since we’re already over the word limit, I prefer not to mention it in the paper, and I plan to merge it into the Stilman agent as a configuration option anyway. I’ve also added a guide to the docs on creating custom agents.

Regarding your first point about citations to related work, I’ll need to follow up with you next week after discussing with the other authors. Thanks for your patience.
