### 2025-09-01 fa41fdf
add citations
