+++ paper.bib
@@ -39,17 +39,84 @@
 }
 
 @article{stilman_2005,
-  title = {Navigation among Movable Obstacles: Real-Time Reasoning in Complex Environments},
-  author = {Stilman, M. and Kuffner, J. J.},
-  journal = {International Journal of Humanoid Robotics},
-  volume = {2},
-  number = {4},
-  year = {2005},
-  pages = {479--503},
-  issn = {0219-8436},
-  doi = {10.1142/S0219843605000545},
-  url = {http://www.worldscientific.com/doi/abs/10.1142/S0219843605000545},
-  urldate = {2018-02-28},
-  abstract = {In this paper, we address the problem of Navigation Among Movable Obstacles (NAMO): a practical extension to navigation for humanoids and other dexterous mobile robots. The robot is permitted to reconfigure the environment by moving obstacles and clearing free space for a path. This paper presents a resolution complete planner for a subclass of NAMO problems. Our planner takes advantage of the navigational structure through state-space decomposition and heuristic search. The planning complexity is reduced to the difficulty of the specific navigation task, rather than the dimensionality of the multi-object domain. We demonstrate real-time results for spaces that contain large numbers of movable obstacles. We also present a practical framework for single-agent search that can be used in algorithmic reasoning about this domain.},
+  title      = {Navigation among Movable Obstacles: Real-Time Reasoning in Complex Environments},
+  author     = {Stilman, M. and Kuffner, J. J.},
+  journal    = {International Journal of Humanoid Robotics},
+  volume     = {2},
+  number     = {4},
+  year       = {2005},
+  pages      = {479--503},
+  issn       = {0219-8436},
+  doi        = {10.1142/S0219843605000545},
+  url        = {http://www.worldscientific.com/doi/abs/10.1142/S0219843605000545},
+  urldate    = {2018-02-28},
+  abstract   = {In this paper, we address the problem of Navigation Among Movable Obstacles (NAMO): a practical extension to navigation for humanoids and other dexterous mobile robots. The robot is permitted to reconfigure the environment by moving obstacles and clearing free space for a path. This paper presents a resolution complete planner for a subclass of NAMO problems. Our planner takes advantage of the navigational structure through state-space decomposition and heuristic search. The planning complexity is reduced to the difficulty of the specific navigation task, rather than the dimensionality of the multi-object domain. We demonstrate real-time results for spaces that contain large numbers of movable obstacles. We also present a practical framework for single-agent search that can be used in algorithmic reasoning about this domain.},
   shorttitle = {Navigation among Movable Obstacles}
+}
+
+@inproceedings{wu_and_levihn,
+  author    = {Hai-Ning Wu and Levihn, Martin and Stilman, Mike},
+  booktitle = {2010 IEEE/RSJ International Conference on Intelligent Robots and Systems},
+  title     = {Navigation Among Movable Obstacles in unknown environments},
+  year      = {2010},
+  volume    = {},
+  number    = {},
+  pages     = {1433-1438},
+  keywords  = {Heuristic algorithms;Optimization;Planning;Navigation;Robot sensing systems;Robot kinematics},
+  doi       = {10.1109/IROS.2010.5649744}
+}
+
+@inproceedings{levihn_2014,
+  author    = {Levihn, Martin and Stilman, Mike and Christensen, Henrik},
+  booktitle = {2014 IEEE-RAS International Conference on Humanoid Robots},
+  title     = {Locally optimal navigation among movable obstacles in unknown environments},
+  year      = {2014},
+  volume    = {},
+  number    = {},
+  pages     = {86-91},
[…]
