### 2025-09-03 1695a86
cite the collision detection algorithm
