+++ paper.bib
@@ -120,3 +120,19 @@
   hal_id      = {hal-04250608},
   hal_version = {v1}
 }
+
+@inbook{Jiménez1998,
+  author    = {Jim{\'e}nez, P.
+               and Thomas, F.
+               and Torras, C.},
+  editor    = {Laumond, J. -P.},
+  title     = {Collision detection algorithms for motion planning},
+  booktitle = {Robot Motion Planning and Control},
+  year      = {1998},
+  publisher = {Springer Berlin Heidelberg},
+  address   = {Berlin, Heidelberg},
+  pages     = {305--343},
+  isbn      = {978-3-540-40917-5},
+  doi       = {10.1007/BFb0036075},
+  url       = {https://doi.org/10.1007/BFb0036075}
+}
