### 2025-09-03 Chroma-CITI/namosim#10 @david-wb
> No need to delete experimental code! You can keep it with a disclaimer that it is experimental, and just not mention it in the paper.

I might bring it back in a new branch but the JOSS paper is our priority at the moment.
