### 2025-09-03 Chroma-CITI/namosim#10 @david-wb
@sea-bass I have addressed each of your four points above:

* New citations to related work are included in the statement of need.
* `ROS2 -> ROS 2 is` done.
* A high-level architecture diagram has been added.
* N/A

Please take a look and let me know if you think anything else if needed.
