+++ paper.bib
@@ -136,3 +136,22 @@
   doi       = {10.1007/BFb0036075},
   url       = {https://doi.org/10.1007/BFb0036075}
 }
+
+@inproceedings{nav2,
+  author    = {Macenski, Steven and Martin, Francisco and White, Ruffin and Ginés Clavero, Jonatan},
+  title     = {The Marathon 2: A Navigation System},
+  booktitle = {2020 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  year      = {2020}
+}
+
+@inproceedings{Koenig2004,
+  author = {Koenig, Nathan and Howard, Andrew},
+  title = {Design and Use Paradigms for Gazebo, an Open-Source Multi-Robot Simulator},
+  booktitle = {IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  year = {2004},
+  volume = {3},
+  pages = {2149--2154},
+  address = {Sendai, Japan},
+  month = sep,
+  doi = {10.1109/IROS.2004.1389727}
+}
