+++ paper.bib
@@ -145,13 +145,13 @@
 }
 
 @inproceedings{Koenig2004,
-  author = {Koenig, Nathan and Howard, Andrew},
-  title = {Design and Use Paradigms for Gazebo, an Open-Source Multi-Robot Simulator},
+  author    = {Koenig, Nathan and Howard, Andrew},
+  title     = {Design and Use Paradigms for Gazebo, an Open-Source Multi-Robot Simulator},
   booktitle = {IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
-  year = {2004},
-  volume = {3},
-  pages = {2149--2154},
-  address = {Sendai, Japan},
-  month = sep,
-  doi = {10.1109/IROS.2004.1389727}
+  year      = {2004},
+  volume    = {3},
+  pages     = {2149--2154},
+  address   = {Sendai, Japan},
+  month     = sep,
+  doi       = {10.1109/IROS.2004.1389727}
 }
