+++ paper.bib
@@ -155,3 +155,23 @@
   month     = sep,
   doi       = {10.1109/IROS.2004.1389727}
 }
+
+@article{rviz,
+  author     = {Kam, Hyeong Ryeol and Lee, Sung-Ho and Park, Taejung and Kim, Chang-Hun},
+  title      = {RViz: a toolkit for real domain data visualization},
+  year       = {2015},
+  issue_date = {October   2015},
+  publisher  = {Kluwer Academic Publishers},
+  address    = {USA},
+  volume     = {60},
+  number     = {2},
+  issn       = {1018-4864},
+  url        = {https://doi.org/10.1007/s11235-015-0034-5},
+  doi        = {10.1007/s11235-015-0034-5},
+  abstract   = {In computational science and computer graphics, there is a strong requirement to represent and visualize information in the real domain, and many visualization data structures and algorithms have been proposed to achieve this aim. Unfortunately, the dataflow model that is often selected to address this issue in visualization systems is not flexible enough to visualize newly invented data structures and algorithms because this scheme can accept only specific data structures. To address this problem, we propose a new visualization tool, RViz, which is independent of the input information data structures. Since there is no requirement for additional efforts to manage the flow networks and the interface to abstracted information is simple in RViz, any scientific information visualization algorithms are easier to implement than the dataflow model. In this paper, we provide case studies in which we have successfully implemented new data structures and related algorithms using RViz, including geometry synthesis, distance field representation, and implicit surface reconstruction. Through these cases, we show how RViz helps users visualize and understand any hidden insights in input information.},
+  journal    = {Telecommun. Syst.},
+  month      = oct,
+  pages      = {337–345},
+  numpages   = {9},
+  keywords   = {Decorator pattern, Lightweight framework, Spatial data visualization, Visualization toolkit}
+}
