### 2025-09-05 Chroma-CITI/namosim#10 @sea-bass
Thanks @david-wb! I just have one remaining suggestion.

It's great that you've added more academic references, but I think for JOSS papers you also want to compare against (and cite) competitive software packages as well. In other words, how does NAMOSIM complement or improve upon other existing software tools for multi-agent planning for NAMO applications?
