### 2025-09-05 Chroma-CITI/namosim#10 @david-wb
> Thanks [@david-wb](https://github.com/david-wb)! I just have one remaining suggestion.
>
> It's great that you've added more academic references, but I think for JOSS papers you also want to compare against (and cite) competitive software packages as well. In other words, how does NAMOSIM complement or improve upon other existing software tools for multi-agent planning for NAMO applications?

@sea-bass That's just the thing. There don't seem to be any easily-reusable software packages designed for this problem. I could contrast with multi-robot MAPF packages, but those don't have the movable-obstacle component. The closest thing I can find to multi-agent NAMO is [this work](https://openai.com/index/emergent-tool-use/) from open-ai, but it is neither designed for navigation nor even for robotics. I'll see about mentioning these in the paper. If you're aware of any other relevant packages, I'd be happy to take a look at them.
