### 2025-09-05 Chroma-CITI/namosim#10 @sea-bass
Ah, if that's the case then you can include in the statement of need that there was nothing, so as far as you know it serves a unique need!
