+++ paper.bib
@@ -175,3 +175,25 @@
   numpages   = {9},
   keywords   = {Decorator pattern, Lightweight framework, Spatial data visualization, Visualization toolkit}
 }
+
+@misc{baker2020emergenttoolusemultiagent,
+  title         = {Emergent Tool Use From Multi-Agent Autocurricula},
+  author        = {Bowen Baker and Ingmar Kanitscheider and Todor Markov and Yi Wu and Glenn Powell and Bob McGrew and Igor Mordatch},
+  year          = {2020},
+  eprint        = {1909.07528},
+  archiveprefix = {arXiv},
+  primaryclass  = {cs.LG},
+  url           = {https://arxiv.org/abs/1909.07528}
+}
+
+@article{ros2,
+  author  = {Steven Macenski and Tully Foote and Brian Gerkey and Chris Lalancette and William Woodall},
+  title   = {Robot Operating System 2: Design, architecture, and uses in the wild},
+  journal = {Science Robotics},
+  volume  = {7},
+  number  = {66},
+  pages   = {eabm6074},
+  year    = {2022},
+  doi     = {10.1126/scirobotics.abm6074},
+  url     = {https://www.science.org/doi/abs/10.1126/scirobotics.abm6074}
+}
