+++ paper.md
@@ -38,7 +38,7 @@ bibliography: paper.bib
 
 # Summary
 
-**NAMOSIM** is a mobile robot motion planner and simulator designed for the problem of **N**avigation **A**mong **M**ovable **O**bstacles (NAMO). The planner simulates robots navigating in 2D polygonal environments where certain obstacles can be grasped and relocated to enable robots to reach their goals. NAMOSIM thus extends the classic navigation problem with a layer of interactivity, posing interesting research questions while remaining well-defined and amenable to various algorithmic approaches. NAMOSIM is intended for researchers and developers working on robot navigation in dynamic environments, particularly where physical interaction is necessary.
+**NAMOSIM** is a mobile robot motion planner and simulator designed for the problem of **N**avigation **A**mong **M**ovable **O**bstacles (NAMO). The planner simulates robots navigating in 2D polygonal environments where certain obstacles can be grasped and relocated to enable robots to reach their goals. NAMOSIM extends the classic navigation problem with a layer of interactivity, posing interesting research questions while remaining well-defined and amenable to various algorithmic approaches. NAMOSIM is intended for researchers and developers working on robot navigation in dynamic environments, particularly where physical interaction is necessary.
 
 NAMOSIM supports the development of custom NAMO algorithms using a modular agent-based architecture. It includes a baseline agent implementing Stilman's NAMO algorithm [@stilman_2005] and incorporating a communication-free coordination strategy for multi-robot scenarios [@renault_2024_iros]. A variety of other agent behaviors are implemented, and new agents utilizing alternative approaches can be integrated into the planner by implementing the **Agent** base class. NAMOSIM thus supports reproducible research in single and multi-robot NAMO algorithms.
 
@@ -48,11 +48,11 @@ NAMOSIM is packaged as a ROS 2 package for easy integration into robotics projec
 
 Many interesting applications in autonomous mobile robotics involve physical interaction with the environment as well as social coordination with other agents.
 However, global navigation planners typically assume static environments, leaving complex behaviors to be managed by separate software components, complicating implementation.
-Ideally, motion planners should reason about physical and social interactions and adapt to changing conditions. NAMOSIM addresses this challenge by providing an open-source planner and simulation environment designed for single and multi-robot NAMO algorithms.
+Ideally, motion planners should be able to reason about physical and social interactions and adapt to changing conditions. NAMOSIM addresses this challenge by providing an open-source planner and simulation environment designed for single and multi-robot NAMO algorithms, which require not only navigation but also reasoning about which obstacles to move, where to move them, and how to coordinate with other agents.
 
-While prior works on NAMO [@stilman_2005; @wu_and_levihn; @levihn_2013; @levihn_2014; @scholz_2016; @zhang_2023] have introduced new methodologies, an open-source package dedicated to developing NAMO algorithms and compatible with real-robot platforms is lacking.
+While there are several prior works on NAMO [@wu_and_levihn; @levihn_2013; @levihn_2014; @scholz_2016; @zhang_2023], they lack multi-robot capability and do not provide reusable open-source tools. There is also a large body of work on multi-agent path finding (MAPF), but this does not consider movable obstacles. Another innovative related work [@baker2020emergenttoolusemultiagent] involved both multiple agents and movable obstacles, but was not designed for navigation nor for robotics applications. NAMOSIM fills this gap by providing a general-purpose software package for single and multi-robot NAMO. 
 
[…]
