+++ paper.md
@@ -42,7 +42,7 @@ bibliography: paper.bib
 
 NAMOSIM supports the development of custom NAMO algorithms using a modular agent-based architecture. It includes a baseline agent implementing Stilman's NAMO algorithm [@stilman_2005] and incorporating a communication-free coordination strategy for multi-robot scenarios [@renault_2024_iros]. A variety of other agent behaviors are implemented, and new agents utilizing alternative approaches can be integrated into the planner by implementing the **Agent** base class. NAMOSIM thus supports reproducible research in single and multi-robot NAMO algorithms.
 
-NAMOSIM is packaged as a ROS 2 package for easy integration into robotics projects but can also be used as a standalone Python module. Simulations are displayed in a Tkinter window and ROS 2 messages are published for more-detailed visualization in RViz [@rviz]. Several prebuilt scenarios for testing and benchmarking are included. These are stored as SVG files, allowing for convenient creation of custom scenarios using a free SVG editor such as Inkscape.
+NAMOSIM is packaged as a ROS 2 [@ros2] package to facilite integration with commonly-used robotics tools such as Nav2 [@nav2] and Gazebo [@Koenig2004], but it can also be used as a standalone Python module. Simulations are displayed in a Tkinter window and ROS 2 messages are published for more-detailed visualization in RViz [@rviz]. Several prebuilt scenarios for testing and benchmarking are included. These are stored as SVG files, allowing for convenient creation of custom scenarios using a free SVG editor such as Inkscape.
 
 # Statement of Need
 
@@ -50,9 +50,7 @@ Many interesting applications in autonomous mobile robotics involve physical int
 However, global navigation planners typically assume static environments, leaving complex behaviors to be managed by separate software components, complicating implementation.
 Ideally, motion planners should be able to reason about physical and social interactions and adapt to changing conditions. NAMOSIM addresses this challenge by providing an open-source planner and simulation environment designed for single and multi-robot NAMO algorithms, which require not only navigation but also reasoning about which obstacles to move, where to move them, and how to coordinate with other agents.
 
-While there are several prior works on NAMO [@wu_and_levihn; @levihn_2013; @levihn_2014; @scholz_2016; @zhang_2023], they lack multi-robot capability and do not provide reusable open-source tools. There is also a large body of work on multi-agent path finding (MAPF), but this does not consider movable obstacles. Another innovative related work [@baker2020emergenttoolusemultiagent] involved both multiple agents and movable obstacles, but was not designed for navigation nor for robotics applications. NAMOSIM fills this gap by providing a general-purpose software package for single and multi-robot NAMO. 
-
-NAMOSIM supports reproducible research and real-world deployment via compatibility with the ROS [@ros2] ecosystem, facilitating integration with commonly-used packages such as Nav2 [@nav2] and Gazebo [@Koenig2004]. By providing a versatile tool for developing NAMO algorithms, NAMOSIM supports the robotics community to create more capable and adaptive robotic systems.
+While there are several prior works on NAMO [@wu_and_levihn; @levihn_2013; @levihn_2014; @scholz_2016; @zhang_2023], they lack multi-robot capability and do not provide reusable open-source tools designed for real-world deployment on physical robots. There is also a large body of work on multi-agent path finding (MAPF), but this does not consider movable obstacles. Another innovative related work [@baker2020emergenttoolusemultiagent] involved both multiple agents and movable obstacles, but was not designed for navigation nor for robotics applications. NAMOSIM fills this gap by providing a versatile software package for experimenting with single and multi-robot NAMO algorithms, supporting the robotics community to create more capable and adapative robotic systems.
 
 # Major Features
 
