### 2025-09-08 Chroma-CITI/namosim#10 @david-wb
@sea-bass Thanks, that part is done. Please have another look when you get a chance.
