### 2025-09-08 #8816 @boisgera
> [@boisgera](https://github.com/boisgera) the author has satisfied all my review comments and I believe the submission meets the bar for publication to JOSS. It's an approve from me!
>
> (Thanks [@david-wb](https://github.com/david-wb)!)

Great, thanks a lot for your work on this!

I notice that you didn't check the "State of the field" item [in your list](https://github.com/openjournals/joss-reviews/issues/8816#issuecomment-3190874047). Is it on purpose?
