### 2025-09-10 #8816 @david-wb
@SteveMacenski ROS 2 is cited. The editorial-bot commands don't work for me, but you can see the updated pdf [here](https://github.com/Chroma-CITI/namosim/blob/humble/paper.pdf).
