### 2025-09-17 #8816 @boisgera
Hi @david-wb,

You will find [above](https://github.com/openjournals/joss-reviews/issues/8816#issuecomment-3303380697) the list of remaining tasks. Please make sure that when your archive title, list of authors, license and version tag match exactly the ones of your JOSS submission.

I have no comment/correction to suggest for your paper at this stage, except for the references: could you please check if the [suggestions made by our editorialbot in ❌ MISSING DOIs](https://github.com/openjournals/joss-reviews/issues/8816#issuecomment-3303348626) are valid and update the bibliography accordingly?

Best regards,

Sébastien
