+++ paper.bib
@@ -1,15 +1,15 @@
 @inproceedings{renault_2024_iros,
-  title     = {Multi-Robot Navigation among Movable Obstacles: Implicit Coordination to Deal with Conflicts and Deadlocks},
   author    = {Renault, Benoit and Saraydaryan, Jacques and Brown, David and Simonin, Olivier},
-  booktitle = {IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
-  address   = {Abu Dhabi, United Arab Emirates},
+  booktitle = {2024 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  title     = {Multi-Robot Navigation Among Movable Obstacles: Implicit Coordination to Deal with Conflicts and Deadlocks},
   year      = {2024},
-  month     = {October},
-  publisher = {IEEE},
-  pages     = {1--7},
-  url       = {https://hal.science/hal-04705395},
-  pdf       = {https://hal.science/hal-04705395v1/file/IROS24_1134_FI.pdf}
+  volume    = {},
+  number    = {},
+  pages     = {3505-3511},
+  keywords  = {Machine learning algorithms;Costs;Navigation;Robot kinematics;Machine learning;System recovery;Benchmark testing;Multi-robot systems;Intelligent robots},
+  doi       = {10.1109/IROS58592.2024.10802092}
 }
+
 
 @inproceedings{renault_2020_iros,
   title     = {Modeling a Social Placement Cost to Extend Navigation Among Movable Obstacles (NAMO) Algorithms},
@@ -93,8 +93,10 @@
   address   = {Berlin, Heidelberg},
   pages     = {19--35},
   abstract  = {In this paper we present the first decision theoretic planner for the problem of Navigation Among Movable Obstacles (NAMO). While efficient planners for NAMO exist, they are challenging to implement in practice due to the inherent uncertainty in both perception and control of real robots. Generalizing existing NAMO planners to nondeterministic domains is particularly difficult due to the sensitivity of MDP methods to task dimensionality. Our work addresses this challenge by combining ideas from Hierarchical Reinforcement Learning with Monte Carlo Tree Search, and results in an algorithm that can be used for fast online planning in uncertain environments. We evaluate our algorithm in simulation, and provide a theoretical argument for our results which suggest linear time complexity in the number of obstacles for typical environments.},
-  isbn      = {978-3-642-36279-8}
+  isbn      = {978-3-642-36279-8},
+  doi       = {10.1007/978-3-642-36279-8_2}
 }
+
 
 @inproceedings{scholz_2016,
   author    = {Scholz, Jonathan and Jindal, Nehchal and Levihn, Martin and Isbell, Charles L. and Christensen, Henrik I.},
@@ -109,17 +111,17 @@
 }
 
 @inproceedings{zhang_2023,
-  title       = {{Navigation among movable obstacles using machine learning based total time cost optimization}},
-  author      = {Zhang, Kai and Lucet, Eric and Alexandre Dit Sandretto, Julien and Filliat, David},
-  url         = {https://ip-paris.hal.science/hal-04250608},
-  booktitle   = {{IROS 2023 - IEEE/RSJ International Conference on Intelligent Robots and Systems}},
-  address     = {Detroit (MI), United States},
-  year        = {2023},
-  month       = Oct,
-  pdf         = {https://ip-paris.hal.science/hal-04250608v1/file/namo_iros2023_preprint.pdf},
-  hal_id      = {hal-04250608},
-  hal_version = {v1}
[…]
