+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package namosim
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Forthcoming
+-----------
+
 0.0.3 (2025-09-08)
 ------------------
 
