### 2025-09-24 0291757
update citations
