### 2025-09-26 16e74f0
Update docs/source/installation.rst
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
